Refactor MessageProcessor and some cleanup.
