+++ crossdc-consumer/src/main/java/org/apache/solr/crossdc/ResubmitBackoffPolicy.java
@@ -1,0 +1,7 @@
+package org.apache.solr.crossdc;
+
+import org.apache.solr.crossdc.common.MirroredSolrRequest;
+
+public interface ResubmitBackoffPolicy {
+  long getBackoffTimeMs(MirroredSolrRequest resubmitRequest);
+}
